IPv4_Sockets: formatting
