+++ test/unit-test/FreeRTOS_IPv4_Sockets/FreeRTOS_IPv4_Sockets_stubs.c
@@ -1,0 +1,66 @@
+/*
+ * FreeRTOS+TCP <DEVELOPMENT BRANCH>
+ * Copyright (C) 2022 Amazon.com, Inc. or its affiliates.  All Rights Reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
+ * http://aws.amazon.com/freertos
+ * http://www.FreeRTOS.org
+ */
+
+
+/* Include Unity header */
+#include <unity.h>
+
+/* Include standard libraries */
+#include <stdlib.h>
+#include <string.h>
+#include <stdint.h>
+#include "FreeRTOS.h"
+#include "task.h"
+#include "list.h"
+
+#include "FreeRTOS_IP.h"
+#include "FreeRTOS_IP_Private.h"
+
+/* =========================== EXTERN VARIABLES =========================== */
+
+char cStubNtoaString[ 128 ];
+
+/* ======================= Stub Callback Functions ======================== */
+
+void setUpStub( void )
+{
+    memset( cStubNtoaString, 0, sizeof( cStubNtoaString ) );
+}
+
+static void prvSetString_FreeRTOS_inet_ntoa( char * pcBuffer )
+{
+    strcpy( cStubNtoaString, pcBuffer );
+}
+
+static const char * pucStub_FreeRTOS_inet_ntoa( uint32_t ulIPAddress,
+                                                char * pcBuffer,
+                                                int numCall )
+{
+    strcpy( pcBuffer, cStubNtoaString );
+
+    return pcBuffer;
+}
